+++ Assets/Digigrams.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc68e7b718eb98b4b8fef30989e84e05
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
