+++ Assets/Digigrams/Digitest.asset
@@ -1,0 +1,18 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2f5420b9ceb2b6d4e81e83d5d9af65d7, type: 3}
+  m_Name: Digitest
+  m_EditorClassIdentifier: 
+  StaticObjects: {fileID: 0}
+  Texture: {fileID: 0}
+  TextureSize: {x: 1080, y: 1080}
+  BitsPerChannel: 16
