+++ Assets/Digigrams/Digitest.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f9456cc82baab3a47b66b59c8cb91504
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
