+++ Assets/Scripts/Digigram.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2f7e8ff344fb86e49bae800bbfba5965
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
