+++ Assets/Scripts/Digigrammer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 80c6cd77470a1df45b86a59afc3bdea2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
