+++ Assets/Scripts/Scheduler/Class1.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7c2d078af7116fa46ab0253c38a721e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
